requirements.txt
streamlit
pandas
psycopg2-binary
plotly
openpyxl
netmiko
sshtunnel
paramiko==2.12.0
